renamed UniverseDeclareAnnotationMirror to UniverseAnnotationMirrorHolder
